Add support for google closure library
When the closure library is installed by npm, it is named "google-closure-library", however it is named "closure-library" in build.py. When both blockly and closure are installed by npm, they are in the same folder so theoretically theoretically blockly should be able to build. Formatting and indentation might be a bit off as I edited this using the github editor. 

This ```python
search_paths = calcdeps.ExpandDirectories(
      ["core", os.path.join(os.path.pardir, "closure-library")])
``` might need further revision to account for the additional naming. Maybe a try statement?