Load builtin assets as ArrayBuffer, calculate IDs
The `BuiltinHelper` now uses `arraybuffer-loader` to load assets as
`ArrayBuffer` instances instead of binary strings. Converting a binary
string to a typed array isn't as straightforward as one might expect.

One side effect of this change is that these builtin assets are now
stored as `ArrayBuffer` instead of `Buffer` instances.

Also, if a built-in asset has no ID at start time, an ID will be
calculated. This makes it easy to give the files human-readable names
like "defaultBitmap.png" and makes updating the files easier.