Set `src/.eslintrc.js` to root as well
The code in `src/` is intended for use in both Node and on the web, but
will be processed by Webpack which emulates a subset of Node's features
when targeting the web. This change separates the linting rules for
`src/` from the root of the project (which is intended for Node
directly) and enforces Node-like linting rules without actually
including Node globals. If we end up needing any of the Node globals
provided by Webpack, we can add them explicitly here.