Fix color button turning white when it shouldn't
If you select a group with "mixed" gradient types, then color and color2
will both be MIXED. If you then select a shape with a solid color, it
won't reset color2, which will remain MIXED. We only want the background
to appear white if color2 is MIXED *and* the gradient type actually uses
color2. This is already done in the component itself, which only shows
the "mixed" icon if the gradient type is non-solid, but
colorToBackground left out this gradient type check.