Replace Stage <div> with React.Fragment
The Stage component will sometimes have a color picker background,
which darkens the entire screen when the color picker loupe is active.
This color picker background element should appear in the same level as
the stage's wrapper, not inside of it. Previously, this was done by
wrapping both the stage's wrapper and color picker background in a div.
Now, they're both wrapped in a React.Fragment to avoid a completely
unnecessary div wrapper.