fix(fetchtool, webhelper): handle asset requests with a 404 response
FetchTool should resolve null for the data request if the request gets a 404 response. WebHelper
should properly handle this case and resolve null for the asset (instead of an asset skeleton with
null 'data' field) if the asset is nowhere to be found. WebHelper should also report errors
properly.

BREAKING CHANGE: The WebHelper was not previously resolving with null assets even though this was
the documented intended behavior of ScratchStorage.load

builds on and closes #363 (thanks @AndreyNikitin!), closes #156, might be related to #367