ci: use DETECT_CHROMEDRIVER_VERSION
This environment variable tells `node-chromedriver` to check the version
of `chrome` and install a matching version of `chromedriver`. This also
means that `chrome` must be installed before `node-chromedriver`, so
I've rearranged the steps in `setup` and `build-test-no-cache` to ensure
that. Installing a system-wide `chromedriver` with `browser-tools` is
not necessary when using `node-chromedriver` since the system-wide copy
gets ignored.