Fix JummBox json importing
Technically this isn't a complete fix.

The problem that this commit deals with is that early versions of
UltraBox added two chip waves with names that already were being used in
JummBox ("trumpet" and "flute"). These were eventually renamed, though
code added for backwards compatibility with those old files didn't
address the ambiguity.

This commit resolves the ambiguity simply by mapping the "trumpet" and
"flute" chip waves to the JummBox ones, instead of the PandorasBox ones,
if these are the only names from this early set of samples seen
(conversely, if one of the non-ambiguous names is seen, then the
PandorasBox ones are used for "trumpet" and "flute" as well).

In the future, the ambiguity should be resolved by the user at import
time (see issue #68).

This should also avoid errors in general when looking at what chip wave
should be used when importing from json, though it may assign incorrect
values for that.