Lily/SoundExpanded: fix "change project volume" block (#1423)
It used to just always set the volume to 0. This fixes it.

This also makes the set volume block be clamped to 0-100 instead of
being wrapped around 0-199, which I think makes more sense and is
consistent with the vanilla set volume block. Though I'm not sure if it
wrapping around 199 instead of 100 (or 99) was a mistake or not, or if
it'll break projects.