Rename the "Scratch" font to something that isn't a trademark
Background:
The font called "Scratch" in Scratch 2 is actually "Boys on mopeds".
For license reasons we use a different font based on "Life Savers" that looks similar.
The creator called it "Scratch Savers" but that has the same problem as "Scratch".

If you have a better idea forwhat we call this font then let me know