rant in desc, but revert this for now
I originally was told about a bug in paper.js by Cryptid, and this was meant to fix that. Apparently, this only fixes the bug in render time and does not fix the real problem. After this commit, people in the TW server called me lazy and said I "fucked it up" when it was easy for trying to fix a bug. I did not know how to actually fix the issue so I've left this in PM since then. Even today, no one there has told me how to fix it once even after I asked so I'm going to figure it out myself.

This reverts commit 614e5741bc1ff3c1aa88af6d5df67b02f7573c7b.