Remove CSS sanitization in favor of the sandbox
I believe that attempting to parse CSS and enumerate all the possible
ways for CSS to make an external request is not a good path to go down.
There are too many possible bugs from parser bugs, new CSS features,
weird escaping rules, invalid syntax being interpreted differently, ...

Instead we will rely on the sandbox which lets the browser do the hard
part.

The CSS sanitization code that this commit removes has known weaknesses,
so it is an ineffective layer of protection. Removing it does not
negatively impact security at all as it already does not provide any
security.

css-tree is also a reasonably large library which is always nice to
remove.