Add isMetricsReady for use in bounds computation
Bounds computations don't need to get a fully updated copy of
the skin's texture. It just needs to know the size and rotation
center, which is what this method can check for.

Default implementation defaults to !!getTexture([100, 100]) for
compatibility with quirks relying on the old behavior. The
getTexture() functions in scratch-render don't fall prey to issues
such as https://github.com/TurboWarp/extensions/issues/2207
so this default implementation is fine.
(TextBubbleSkin almost does, but its drawables are non-interactive)

Closes https://github.com/TurboWarp/scratch-render/issues/16