text.js -- Add array of regex matches block (#2306)
Added modified versions of `split` and `matchRegex` that both return an
array instead of a specific item, which can be useful in more complex
projects. I also moved `testRegex` so all of the match blocks could be
together in the palette.


Pretty quick, so there isn't that much polish. It was mostly an easy
copy-and-paste job and things are still left in.

---------

Co-authored-by: DangoCat[bot] <dangocat@users.noreply.github.com>
Co-authored-by: SharkPool <139097378+SharkPool-SP@users.noreply.github.com>